requirement.txt
transformers==4.11.3
spacy==3.0.0
pytorch-crf==0.7.2